+++ back-end/src/main/java/com/apple/shop/cart/CartRepository.java
@@ -3,6 +3,9 @@ package com.apple.shop.cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface CartRepository extends JpaRepository<Cart, Long> {
+    List<Cart> findByMemberId(String memberId);
 }